+++ content/tests.py
@@ -24,7 +24,7 @@ from users.models import User, RegUser, Profile
 
 # content function imports
 from .models import award_challenge_win, QuizQuestion, FreeTextQuestion, PictureQuestion, QuestionOption, \
-    award_first_goal, CustomNotification
+    award_first_goal, CustomNotification, ParticipantPicture
 
 # content model imports
 from .models import Badge, BadgeSettings, UserBadge
@@ -336,6 +336,64 @@ class ChallengeParticipantIntegrationAPI(APITestCase):
 
         self.assertEqual(response.data['id'], challenge_second.id, "Unexpected challenge returned.")
 
+class TestParticipantPicture(APITestCase):
+
+    def test_adding_caption_to_participant_picture_that_exists(self):
+        user = create_test_regular_user('anon')
+
+        challenge_first = Challenge.objects.create(
+            name='First Challenge',
+            activation_date=timezone.now() + timedelta(days=-7),
+            deactivation_date=timezone.now() + timedelta(days=7,),
+            type=2
+        )
+        challenge_first.publish()
+        challenge_first.save()
+
+        participant = Participant.objects.create(user=user,
+                                   challenge=challenge_first,
+                                   date_created=timezone.now(),
+                                   date_completed=timezone.now())
+
+        question = PictureQuestion.objects.create(challenge=challenge_first,
+                                                  text="A Question?")
+
+        ParticipantPicture.objects.create(participant=participant,
+                                          question=question)
+
+
+        self.client.force_authenticate(user=user)
+
+        response = self.client.post(reverse('api:participantpicture-caption', kwargs={'pk' : participant.id}),
+                                    data={'caption':"caption words"})
+        self.assertEquals(response.status_code, status.HTTP_204_NO_CONTENT)
+
+    def test_adding_caption_to_participant_picture_that_does_not_exist(self):
+        user = create_test_regular_user('anon')
+
+        challenge_first = Challenge.objects.create(
+            name='First Challenge',
+            activation_date=timezone.now() + timedelta(days=-7),
+            deactivation_date=timezone.now() + timedelta(days=7,),
+            type=2
+        )
+        challenge_first.publish()
+        challenge_first.save()
+
+        participant = Participant.objects.create(user=user,
+                                   challenge=challenge_first,
+                                   date_created=timezone.now(),
+                                   date_completed=timezone.now())
+
+        question = PictureQuestion.objects.create(challenge=challenge_first,
+                                                  text="A Question?")
+
+
+        self.client.force_authenticate(user=user)
+
+        response = self.client.post(reverse('api:participantpicture-caption', kwargs={'pk' : participant.id}),
+                                    data={'caption':"caption words"})
+        self.assertEquals(response.status_code, status.HTTP_404_NOT_FOUND)
 
 # ==== #
 # Tips #
@@ -784,7 +842,8 @@ class TestGoalAPI(APITestCase):
             "start_date": datetime.utcnow().strftime('%Y-%m-%d'),
             "end_date": datetime.utcnow().strftime('%Y-%m-%d'),
             "target": 1000,
-            "image": None
+            "image": None,
+            "weekly_target": 100
         }
 
         self.client.force_authenticate(user=user)
@@ -805,7 +864,8 @@ class TestGoalAPI(APITestCase):
             "start_date": '2015-11-01',
             "end_date": '2015-11-30',
             "target": 9000,
-            "image": None
+            "image": None,
+            "weekly_target": 1000
         }
 
         self.client.force_authenticate(user=user)
@@ -832,7 +892,8 @@ class TestGoalAPI(APITestCase):
             "start_date": '2015-11-01',
             "end_date": '2015-11-30',
             "target": 9000,
-            "image": None
+            "image": None,
+            "weekly_target": 2500
         }
 
         self.assertEquals(goal.end_date_modified, None, "Unmodified goal end date should be null")
@@ -864,6 +925,7 @@ class TestGoalAPI(APITestCase):
             "start_date": datetime.utcnow().strftime('%Y-%m-%d'),
             "end_date": datetime.utcnow().strftime('%Y-%m-%d'),
             "target": 1000,
+            "weekly_target": 250,
             "image": None,
             "user": user_2.id  # Different user
         }
@@ -956,6 +1018,7 @@ class TestGoalTransactionAPI(APITestCase):
             "start_date": datetime.utcnow().strftime('%Y-%m-%d'),
             "end_date": datetime.utcnow().strftime('%Y-%m-%d'),
             "target": 9000,
+            "weekly_target": 1000,
             "transactions": [{
                 # Duplicate
                 "date": trans.date.isoformat(),
@@ -1005,7 +1068,8 @@ class TestGoalPrototypesAPI(APITestCase):
             'target': 12000,
             'start_date': datetime(2016, 11, 1, tzinfo=timezone.utc).strftime('%Y-%m-%d'),
             'end_date': datetime(2016, 11, 30, tzinfo=timezone.utc).strftime('%Y-%m-%d'),
-            'prototype': proto.id
+            'prototype': proto.id,
+            "weekly_target": 1000
         }
 
         self.client.force_authenticate(user=user)
@@ -1439,7 +1503,8 @@ class TestBadgeAwarding(APITestCase):
             'name': 'Goal 1',
             'target': 10000,
             'start_date': '2016-11-01',
-            'end_date': '2016-11-30'
+            'end_date': '2016-11-30',
+            'weekly_target': 2500
         }
 
         self.client.force_authenticate(user=user)
@@ -1455,7 +1520,8 @@ class TestBadgeAwarding(APITestCase):
             user=user,
             target=100000,
             start_date=now - timedelta(days=30),
-            end_date=now
+            end_date=now,
+            weekly_target=5000
         )
         UserBadge.objects.create(user=user, badge=self.goal_first_created)
 
@@ -1463,7 +1529,8 @@ class TestBadgeAwarding(APITestCase):
             'name': 'Goal 1',
             'target': 10000,
             'start_date': '2016-11-01',
-            'end_date': '2016-11-30'
+            'end_date': '2016-11-30',
+            'weekly_target': 2500
         }
 
         self.client.force_authenticate(user=user)
@@ -1483,7 +1550,8 @@ class TestBadgeAwarding(APITestCase):
             user=user,
             target=10000,
             start_date=now - timedelta(days=30),
-            end_date=now + timedelta(days=30)
+            end_date=now + timedelta(days=30),
+            weekly_target=500
         )
 
         data = [{